fix: Change GET point
